requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.2.12
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.11.0
requests==2.32.5
reportlab==4.4.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
